Fix linting errors: rename xaa-util to xaaUtil, fix unicorn rules
